+++ player/mixer.c
@@ -843,7 +843,7 @@ uint32_t csf_create_stereo_mix(song_t *csf, uint32_t count)
 						channel->flags &= ~(CHN_LOOP_WRAPPED);
 
 					channel->current_sample_data = smp_ptr;
-					if ((int32_t)channel->position >= lookahead_start) {
+					if (channel->position >= lookahead_start) {
 						int32_t samples_to_read = (channel->increment < 0)
 							? ((int32_t)channel->position - lookahead_start)
 							: (channel->loop_end - (int32_t)channel->position);
@@ -856,7 +856,7 @@ uint32_t csf_create_stereo_mix(song_t *csf, uint32_t count)
 						// Interpolate properly after looping
 						smpcount = samples_to_buffer_length(channel->loop_start + MAX_INTERPOLATION_LOOKAHEAD_BUFFER_SIZE - channel->position, channel);
 						channel->current_sample_data = lookahead_ptr + ((channel->loop_end - channel->loop_start) * ((channel->ptr_sample->flags & CHN_STEREO) ? 2 : 1) * ((channel->ptr_sample->flags & CHN_16BIT) ? 2 : 1));
-					} else if (channel->increment >= 0 && pos_dest >= lookahead_start && smpcount > 1) {
+					} else if (channel->increment >= 0 && pos_dest >= (int32_t)lookahead_start && smpcount > 1) {
 						smpcount = samples_to_buffer_length(lookahead_start - channel->position, channel);
 					}
